+++ src/lib/interfaces/IEventSubscriber.ts
@@ -1,4 +1,0 @@
-export interface IEventSubscriber<T extends Record<string, (...args: any[]) => void>> {
-	subscribe<K extends keyof T>(name: K, callback: T[K], priority?: number): () => void;
-	emit<K extends keyof T>(name: K, ...args: Parameters<T[K]>): (ReturnType<T[K]> | Error)[];
-}